Delete older implementation, fixing bugs with server/worker decouple
