--- src/workflows/agent/prompts.ts
@@ -1,141 +1,0 @@
-import { PromptTemplate } from "@langchain/core/prompts";
-
-export function thoughtPromptTemplate() {
-  const templateString = `You are a ReAct (Reasoning and Acting) agent tasked with answering the following query:
-
-<user-query>
-{userQuery}
-</user-query>
-
-Your goal is to reason about the query and decide on the best course of action to answer it accurately.
-
-Instructions:
-1. Analyze the query, previous reasoning steps, and observations.
-2. Decide on the next action: use a tool or provide a final answer.
-3. Respond in the following JSON format:
-
-If you cannot answer the question yet and need to use a tool:
-{{
-    "thought": "Your detailed reasoning about what to do next",
-    "action": {{
-        "name": "Tool name",
-        "reason": "Explanation of why you chose this tool",
-        "input": "JSON object matching to tool input schema"
-    }}
-}}
-
-If you have enough information to answer the query, instead use this format:
-{{
-    "thought": "Your final reasoning process",
-    "answer": "Your comprehensive answer to the query"
-}}
-
-Remember:
-- Be thorough in your reasoning.
-- Use tools when you need more information.
-- Use tools to validate your assumptions and internal knowledge.
-- Be sure to match the tool input schema exactly.
-- Always base your reasoning on the actual observations from tool use.
-- If a tool returns no results or fails, acknowledge this and consider using a different tool or approach.
-- Provide a final answer only when you're confident you have sufficient information.
-- If you cannot find the necessary information after using available tools, admit that you don't have enough information to answer the query confidently.
-- Your internal knowledge may be outdated. The current date is {currentDate}.
-
-You do not need to include any XML tags such as <thought>, <action>, or <observation> in your response, those will be added automatically by the Agent Workflow.
-
-In this thinking step, consider the following information from previous steps:
-
-<previous-steps>
-{previousSteps}
-</previous-steps>
-
-Based on that information, provide your thought process and decide on the next action.
-<available-actions>
-{availableActions}
-</available-actions>
-
-Return your response as a valid JSON object only, with no additional text, explanations, or formatting outside the JSON. Only respond with one singular JSON object, you may not respond with multiple JSON objects.`;
-
-  const prompt = new PromptTemplate({
-    template: templateString,
-    inputVariables: [
-      "userQuery",
-      "currentDate",
-      "previousSteps",
-      "availableActions",
-    ],
-  });
-
-  return prompt;
-}
-
-export function observationPromptTemplate() {
-  const templateString = `You are a ReAct (Reasoning and Acting) agent tasked with answering the following query:
-
-<user-query>
-{userQuery}
-</user-query>
-
-Your goal is to extract insights from the results of your last action and provide a concise observation.
-
-Instructions:
-1. Analyze the query, previous reasoning steps, and observations.
-2. Extract insights from the latest action result.
-3. Respond with a concise observation that summarizes the results of the last action taken.
-
-You do not need to include any XML tags such as <thought>, <action>, or <observation> in your response, those will be added automatically by the Agent Workflow.
-
-In this observation step, consider the following information from previous steps:
-
-<previous-steps>
-{previousSteps}
-</previous-steps>
-
-Provide your observation based on the latest action result:
-<action-result>
-{actionResult}
-</action-result>
-`;
-
-  const prompt = new PromptTemplate({
-    template: templateString,
-    inputVariables: ["userQuery", "previousSteps", "actionResult"],
-  });
-
-  return prompt;
-}
-
-export function compactPromptTemplate() {
-  const templateString = `You are a summarization agent tasked with compacting the context of a ReAct (Reasoning and Acting) agent.
-  
-Your goal is to summarize the provided context, attempting to preserve the most important parts of the context history.
-
-Instructions:
-1. Review the provided context history.
-2. Summarize the context, focusing on preserving key information and recent steps.
-3. Ensure that the most recent parts of the context remain intact.
-
-You do not need to include any XML tags such as <thought>, <action>, or <observation> in your response, those will be added automatically by the Agent Workflow.
-
-For reference, here is the oringinal user question that the agent is trying to answer:
-
-<user-query>
-{userQuery}
-</user-query>
-
-Here is the context history to be compacted:
-
-<context-history>
-{contextHistory}
-</context-history>
-
-Provide a compacted version of the context history, preserving important details and recent steps.
-`;
-
-  const prompt = new PromptTemplate({
-    template: templateString,
-    inputVariables: ["userQuery", "contextHistory"],
-  });
-
-  return prompt;
-}
