--- src/workflows/workflows.ts
@@ -1,2 +1,0 @@
-export * from "./agent/workflow";
-export * from "./agent-entity/workflow";
